+++ Kernel_d/include/CGAL/Kernel_d/Matrix__.h
@@ -221,7 +221,7 @@ void range_initialize(RAIterator first, RAIterator last,
                       std::random_access_iterator_tag) 
 { typedef typename std::iterator_traits<RAIterator>::value_type value_type;
   typedef typename value_type::const_iterator const_iterator;
-  dn_ = last-first;
+  dn_ = static_cast<int>(last-first);
   if (dn_ == 0) { dm_=0; v_=0; return; }
   dm_ = first->dimension(); 
   if (dm_ > 0) {
